// src/app/models/departamentos.ts
export class Departamentos {
    id: number=0;
    create_by: string  = "";
    createDate: string  = "";
    modified_by: string  = "";
    modified_date: string  = "";
    status: string  = "";
    descripcion: string  = "";
    name: string  = "";
    phone: string  = "";
    id_empresa: number=0;
    empresaName:string="";
    
}